refator GhatForm & EditGhat
