+++ src/components/GhatForm/GhatForm.tsx
@@ -4,18 +4,38 @@ import InputWithHelper from 'components/shared/InputWithHelper'
 import Label from 'components/shared/Label'
 import LocationMarker from 'components/shared/LocationMarker'
 import Textarea from 'components/shared/Textarea'
-import React, { FormEvent, useState } from 'react'
+import React, { FormEvent, useEffect, useState } from 'react'
 import toast from 'react-hot-toast'
 import GhatService from 'services/ghats'
 import GoogleMapReact from 'google-map-react'
+import { useQuery, useQueryClient } from 'react-query'
 
-const GhatForm = () => {
-  const [title, setTitle] = useState('')
-  const [description, setDescription] = useState('')
+type GhatFormProp = {
+  ghatId?: string
+}
+
+const GhatForm: React.FC<GhatFormProp> = ({ ghatId }) => {
+  const { data: ghat } = useQuery('getGhatId', async () => {
+    const { data } = await GhatService.getGhatId(ghatId || '')
+    return data
+  })
+  const [title, setTitle] = useState(ghat?.title || '')
+  const [description, setDescription] = useState(ghat?.description || '')
+  const [picture, setPicture] = useState(ghat?.picture || '')
   const [file, setFile] = useState<File | null>(null)
-  const [isDisabled, setIsDisabled] = useState(false)
-  const [lat, setLat] = useState(25.3425829)
-  const [lng, setLng] = useState(82.9702298)
+  const [isLoading, setIsLoading] = useState(false)
+  const [lat, setLat] = useState(ghat?.location?.lat || 25.3425829)
+  const [lng, setLng] = useState(ghat?.location?.lng || 82.9702298)
+
+  const queryClient = useQueryClient()
+
+  useEffect(() => {
+    setTitle(ghat?.title || '')
+    setDescription(ghat?.description || '')
+    setPicture(ghat?.picture || '')
+    setLat(ghat?.location?.lat || 25.3425829)
+    setLng(ghat?.location?.lng || 82.9702298)
+  }, [ghat])
 
   const handleSubmit = (event: FormEvent<HTMLFormElement>) => {
     event.preventDefault()
@@ -24,24 +44,35 @@ const GhatForm = () => {
     formData.append('description', description)
     formData.append('location', JSON.stringify({ lat, lng }))
     if (file) formData.append('picture', file)
-    setIsDisabled(true)
+    setIsLoading(true)
 
-    const addGhat = GhatService.addGhat(formData)
-
-    toast
-      .promise(addGhat, {
-        loading: 'Adding Ghat',
-        success: 'Ghat added successfully.',
-        error: 'Error adding, please try again',
-      })
-      .then(() => {
-        setDescription('')
-        setTitle('')
-        setFile(null)
-      })
-      .finally(() => {
-        setIsDisabled(false)
-      })
+    if (!title) {
+      toast
+        .promise(GhatService.addGhat(formData), {
+          loading: 'Adding Ghat',
+          success: 'Ghat added successfully.',
+          error: 'Error adding, please try again',
+        })
+        .then(() => {
+          setDescription('')
+          setTitle('')
+          setFile(null)
+        })
+        .finally(() => {
+          setIsLoading(false)
+        })
+    } else {
+      toast
+        .promise(GhatService.updateGhatInfo(formData, ghat?._id), {
+          loading: 'Updating...',
+          success: 'updated',
+          error: 'Error updating info',
+        })
+        .then(() => queryClient.invalidateQueries('getGhats'))
+        .finally(() => {
+          setIsLoading(false)
+        })
+    }
   }
 
   return (
@@ -50,7 +81,6 @@ const GhatForm = () => {
         <form
           className="grid grid-cols-1 gap-6"
           action="#"
-          method="post"
           onSubmit={handleSubmit}
         >
           <div className="flex items-center justify-between">
@@ -74,11 +104,7 @@ const GhatForm = () => {
               />
             </div>
             <div>
-              <ImageUpload
-                file={file}
-                // title={ghat?.title}
-                setFile={setFile}
-              />
+              <ImageUpload file={file} setFile={setFile} imgUrl={picture} />
             </div>
           </div>
           <div className="listingSection__wrap">
@@ -106,7 +132,7 @@ const GhatForm = () => {
               </div>
             </div>
           </div>
-          <ButtonPrimary disabled={isDisabled}>Save</ButtonPrimary>
+          <ButtonPrimary loading={isLoading}>Save</ButtonPrimary>
         </form>
       </div>
     </div>
